CR fixes

+++ tools/fw-logger/string-formatter.h
@@ -15,7 +15,7 @@ namespace fw_logger
         string_formatter(std::unordered_map<std::string, std::vector<std::pair<int, std::string>>> enums);
         ~string_formatter(void);
 
-        bool generate_message(const std::string& source, int num_of_params, const uint32_t* params, std::string* dest);
+        bool generate_message(const std::string& source, size_t num_of_params, const uint32_t* params, std::string* dest);
 
     private:
         bool replace_params(const std::string& source, const std::map<std::string, std::string>& exp_replace_map, const std::map<std::string, int>& enum_replace_map, std::string* dest);
